LPD-57281 Move tests to correct path
